Make tokens host specific
If a user runs github-token in two different hosts, the second will fail
as the token name is already registered. Given that token values cannot
be retrieved the best choice seems to be to get a token per host.
By including the node in the token name that is achieved

Closes #1